extract cache manager
