+++ flutter/lib/benchmark/resource_manager.dart
@@ -6,12 +6,12 @@ import 'package:flutter/services.dart';
 import 'package:mlcommons_ios_app/benchmark/configurations_manager.dart';
 
 import 'package:path_provider/path_provider.dart';
-import 'package:wakelock/wakelock.dart';
 import 'package:yaml/yaml.dart';
 
 import 'package:mlcommons_ios_app/benchmark/file_cache_helper_.dart';
 
 import 'archive_cache_helper.dart';
+import 'cache_manager.dart';
 import 'result_manager.dart';
 
 bool isInternetResource(String uri) =>
@@ -35,18 +35,15 @@ class ResourceManager {
 
   final VoidCallback onUpdate;
 
-  Map<String, String> _resourcesMap = {};
   bool _done = false;
-  List<String> _resources = [];
-  double _progress = 0;
+  String _progressString = '0%';
 
   late final String applicationDirectory;
   late final String loadedResourcesDir;
 
   late final List<BatchPreset> _batchPresets;
 
-  late final FileCacheHelper fileCacheHelper;
-  late final ArchiveCacheHelper archiveCacheHelper;
+  late final CacheManager cacheManager;
   late final ConfigurationsManager configurationsManager;
   late final ResultManager resultManager;
 
@@ -55,7 +52,7 @@ class ResourceManager {
   bool get done => _done;
 
   String get progress {
-    return '${(_progress * 100).round()}%';
+    return _progressString;
   }
 
   String get(String uri) {
@@ -67,7 +64,7 @@ class ResourceManager {
       return resourceSystemPath;
     }
 
-    return _resourcesMap[uri]!;
+    return cacheManager.get(uri)!;
   }
 
   Future<bool> isResourceExist(String? uri) async {
@@ -78,6 +75,29 @@ class ResourceManager {
     return path == '' ||
         await File(path).exists() ||
         await Directory(path).exists();
+  }
+
+  void handleResources(List<String> resources, bool purgeOldCache) async {
+    _done = false;
+    onUpdate();
+
+    var internetResources = <String>[];
+    for (final resource in resources) {
+      if (resource.startsWith(_applicationDirectoryPrefix)) continue;
+      if (isInternetResource(resource)) {
+        internetResources.add(resource);
+        continue;
+      }
+      throw 'forbidden path: $resource (only http://, https:// and app:// resources are allowed)';
+    }
+    
+    await cacheManager.cache(internetResources, (double val){ 
+      _progressString = '${(val * 100).round()}%'; 
+      onUpdate();
+      }, purgeOldCache);
+      
+    _done = true;
+    onUpdate();
   }
 
   static Future<String> getApplicationDirectory() async {
@@ -103,119 +123,14 @@ class ResourceManager {
     applicationDirectory = await getApplicationDirectory();
 
     loadedResourcesDir = '$applicationDirectory/$_loadedResourcesDirName';
-    fileCacheHelper = FileCacheHelper(loadedResourcesDir);
-    archiveCacheHelper = ArchiveCacheHelper(fileCacheHelper);
+    await Directory(loadedResourcesDir).create();
+    
+    final fileCacheHelper = FileCacheHelper(loadedResourcesDir);
+    final archiveCacheHelper = ArchiveCacheHelper(fileCacheHelper);
+    cacheManager = CacheManager(loadedResourcesDir, fileCacheHelper, archiveCacheHelper);
+
     configurationsManager = ConfigurationsManager(applicationDirectory);
     resultManager = ResultManager(applicationDirectory);
-
-    await Directory(loadedResourcesDir).create();
-  }
-
-  Future<void> deleteLoadedResources(List<String> nonRemovableResources,
-      [int atLeastDaysOld = 0]) async {
-    final directory = Directory(loadedResourcesDir);
-
-    // can't use await for here because we delete files,
-    // and await for on stream from .list() throws exceptions when file is missing
-    for (final file in await directory.list(recursive: true).toList()) {
-      // skip files in already deleted folders
-      if (!await file.exists()) continue;
-      final relativePath = file.path.substring(loadedResourcesDir.length + 1);
-      var nonRemovable = false;
-      for (var resource in nonRemovableResources) {
-        // relativePath.startsWith(resource): if we want to preserve a folder resource
-        // resource.startsWith(relativePath): if we want to preserve a file resource
-        if (relativePath.startsWith(resource) ||
-            resource.startsWith(relativePath)) {
-          nonRemovable = true;
-          break;
-        }
-      }
-      if (nonRemovable) continue;
-      if (atLeastDaysOld > 0) {
-        var stat = await file.stat();
-        if (DateTime.now().difference(stat.modified).inDays < atLeastDaysOld) {
-          continue;
-        }
-      }
-      await file.delete(recursive: true);
-    }
-  }
-
-  Future<void> purgeOutdatedCache(int atLeastDaysOld) async {
-    var currentResources = <String>[];
-    for (var r in _resourcesMap.values) {
-      if (!r.startsWith(loadedResourcesDir)) continue;
-      currentResources.add(r.substring(loadedResourcesDir.length + 1));
-    }
-    return deleteLoadedResources(currentResources, atLeastDaysOld);
-  }
-
-  bool isResourceAnArchive(String resource) {
-    return resource.endsWith(ArchiveCacheHelper.extension);
-  }
-
-  void handleResources(bool purgeOldCache) async {
-    // disable screen sleep when benchmarks is running
-    await Wakelock.enable();
-
-    final resourcesFromInternet = <String>[];
-    _resourcesMap = {};
-    _progress = 0;
-    _done = false;
-    onUpdate();
-
-    for (final resource in _resources) {
-      if (resource.startsWith(_applicationDirectoryPrefix)) continue;
-      if (_resourcesMap.containsKey(resource)) continue;
-      if (resourcesFromInternet.contains(resource)) continue;
-
-      if (isInternetResource(resource)) {
-        String path;
-        if (isResourceAnArchive(resource)) {
-          path = await archiveCacheHelper.get(resource, false);
-        } else {
-          path = await fileCacheHelper.get(resource, false);
-        }
-        
-        if (path != '') {
-          _resourcesMap[resource] = path;
-          continue;
-        }
-
-        resourcesFromInternet.add(resource);
-
-        continue;
-      }
-
-      throw 'Could not get file path for $resource';
-    }
-    await _handleResourcesFromInternet(resourcesFromInternet);
-
-    _done = true;
-    onUpdate();
-
-    await Wakelock.disable();
-
-    if (purgeOldCache) {
-      const atLeastDaysOld = 30 * 9;
-      await purgeOutdatedCache(atLeastDaysOld);
-    }
-  }
-
-  set resources(List<String> resources) => _resources = resources;
-
-  Future<void> _handleResourcesFromInternet(List<String> resources) async {
-    for (var url in resources) {
-      if (isResourceAnArchive(url)) {
-        _resourcesMap[url] = await archiveCacheHelper.get(url, true);
-      } else {
-        _resourcesMap[url] = await fileCacheHelper.get(url, true);
-      }
-
-      _progress += 1 / resources.length;
-      onUpdate();
-    }
   }
 
   List<BatchPreset> getBatchPresets() {
